variable values contain ranges which are considered in LTL formulas
