+++ extension/src-language-server/stpa/stpa-validator.ts
@@ -165,7 +165,7 @@ export class StpaValidator {
     checkContext(context: Context, accept: ValidationAcceptor): void {
         for (let i = 0; i < context.vars.length; i++) {
             const variable = context.vars[i];
-            const variableValues = variable.ref?.values;
+            const variableValues = variable.ref?.values.map(value => value.name);
             // the value of the variable in the context should be one of the values that are stated in the definition of the variable
             if (!variableValues?.includes(context.values[i])) {
                 accept('error', 'This variable has an invalid value.', { node: context, range: variable.$refNode?.range });
